pokretanje novog nivoa

+++ Project/main_window.py
@@ -71,8 +71,8 @@ class MainWindow(QWidget):
     def center(self):
         screen = QDesktopWidget().screenGeometry()
         size = self.geometry()
-        self.move((screen.width() - size.width()) / 2,
-                  (screen.height() - size.height()) / 2)
+        self.move(int((screen.width() - size.width()) / 2),
+                  int((screen.height() - size.height()) / 2))
 
 
     def setWindowsOptions(self):
